% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{betaUpdate.func}
\alias{betaUpdate.func}
\title{Calculate EM beta updates}
\usage{
\code{betaUpdate.func(parList)}
}
\description{
\code{betaUpdate.func} is a helper function to compute the EM
  update to the beta parameter
}
