% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{betaUpdate2.func}
\alias{betaUpdate2.func}
\title{Calculate EM beta updates at update pars}
\usage{
\code(betaUpdate2.func(parList))
}
\description{
\code{betaUpdate2.func} is a helper function to compute the EM
  update to the beta parameter at update pars
}
