% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{calcDepthRep}
\alias{calcDepthRep}
\title{Calculate replications of depth}
\usage{
\code{calcDepthRep(depth)}
}
\description{
\code{calcDepthRep} is a helper function to calculate the
  replications in the depth vector. Procedure is close to that of
  \code{table}, but maintains the same rounding scheme as other functions.
}
