% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcECDF.R
\name{calcDev}
\alias{calcDev}
\title{Calculate upper and lower deviations at intercept points}
\usage{
\code{calcDev(y, depth, depthRep, slope = 1)}
}
\description{
\code{calcDev} is a helper function to calculate upper and
  lower sums of absolute deviations for quantile regression.
}
