% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSlope.R
\name{calcSlope}
\alias{calcSlope}
\title{Calculate expression-depth slope}
\usage{
\code{calcSlope(counts, depth, nSubCell, nSubGene,
  clusterSlope, nCores, ...)}
}
\description{
\code{calcSlope} is a helper function to calculate the
  expression-depth slope
}
