% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{calcStep.func}
\alias{calcStep.func}
\title{Calculate EM step direction/length}
\usage{
\code{calcStep.func(parList)}
}
\description{
\code{calcStep.func} is a helper function to calculate the
  step direction/length for the modified EM iteration
}
