% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{check_DinoIn}
\alias{check_DinoIn}
\title{Check Dino input}
\usage{
\code{counts, nCores, prec, seed, minNZ, nSubGene, nSubCell, depth,
  slope, returnMeta)}
}
\description{
\code{check_DinoIn} is a helper function to verify the
  correct format of the arguments of \code{Dino}.
}
