% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkCounts}
\alias{checkCounts}
\title{Check \emph{counts} input}
\usage{
\code{checkCounts(counts)}
}
\description{
\code{checkCounts} is a helper function to verify the
  correct format of the \emph{counts} argument
}
