% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkMinNZ}
\alias{checkMinNZ}
\title{Check \emph{minNZ} input}
\usage{
\code{checkMinNZ(minNZ, counts)}
}
\description{
\code{checkMinNZ} is a helper function to verify the
  correct format of the \emph{minNZ} argument
}
