% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkNCores}
\alias{checkNCores}
\title{Check \emph{nCores} input}
\usage{
\code{checkNCores(nCores)}
}
\description{
\code{checkNCores} is a helper function to verify the
  correct format of the \emph{nCores} argument
}
