% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkNSub}
\alias{checkNSub}
\title{Check \emph{nSub} inputs}
\usage{
\code{checkNSub(nSub)}
}
\description{
\code{checkNSub} is a helper function to verify the
  correct format of the \emph{nSub} argument.
}
