% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkNSubCell}
\alias{checkNSubCell}
\title{Check \emph{nSubCell} inputs}
\usage{
\code{checkNSubCell(nSubCell)}
}
\description{
\code{checkNSubCell} is a helper function to verify the
  correct format of the \emph{nSubCell} argument.
}
