% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkNSubGene}
\alias{checkNSubGene}
\title{Check \emph{nSubGene} inputs}
\usage{
\code{checkNSubGene(nSubGene)}
}
\description{
\code{checkNSubGene} is a helper function to verify the
  correct format of the \emph{nSubGene} argument.
}
