% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkPrecDepth}
\alias{checkPrecDepth}
\title{Check \emph{prec} and \emph{depth} inputs}
\usage{
\code{checkPrecDepth(depth, counts)}
}
\description{
\code{checkPrecDepth} is a helper function to verify the
  correct format of the \emph{prec} and \emph{depth} arguments
}
