% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkRetM}
\alias{checkRetM}
\title{Check \emph{returnMeta} inputs}
\usage{
\code{checkRetM(returnMeta)}
}
\description{
\code{returnMeta} is a helper function to verify the
  correct format of the \emph{returnMeta} argument.
}
