% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkSeed}
\alias{checkSeed}
\title{Check \emph{seed} inputs}
\usage{
\code{checkSeed(seed)}
}
\description{
\code{checkSeed} is a helper function to verify the
  correct format of the \emph{seed} argument and adjust the working seed
  for replicability if \emph{seed} is not \code{NULL}.
}
