% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{checkSlope}
\alias{checkSlope}
\title{Check \emph{slope} inputs}
\usage{
\code{checkSlope(slope)}
}
\description{
\code{checkSlope} is a helper function to verify the
  correct format of the \emph{slope} argument.
}
