% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{fPhi.func}
\alias{fPhi.func}
\title{Solve likelihood given dispersion parameter}
\usage{
\code{fPhi.func(phi, yVec, muMat, tMat, yInd, LS)}
}
\description{
\code{fPhi.func} is a helper function to calculate the
  likelihood of the observed data given the fitted mixture model
  for the purpose of finding the maximum likelihood dispersion
  parameter.
}
