% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gamUpdate.func}
\alias{gamUpdate.func}
\title{Calculate EM gamma updates}
\usage{
\code{gamUpdate.func(parList)}
}
\description{
\code{gamUpdate.func} is a helper function to compute the EM
  update to the gamma parameter
}
