% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gamUpdate2.func}
\alias{gamUpdate2.func}
\title{Calculate EM gamma updates at update pars}
\usage{
\code{gamUpdate2.func(parList)}
}
\description{
\code{gamUpdate2.func} is a helper function to compute the EM
  update to the gamma parameter at update pars
}
