% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gBeta.func}
\alias{gBeta.func}
\title{Calculate likelihood gradient in beta}
\usage{
\code(gBeta.func(parList))
}
\description{
\code{gBeta.func} is a helper function to compute the
  likelihood gradient in beta
}
