% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gBeta2.func}
\alias{gBeta2.func}
\title{Calculate likelihood gradient in beta at update pars}
\usage{
\code{gBeta2.func(parList)}
}
\description{
\code{gBeta2.func} is a helper function to compute the
  likelihood gradient in beta at update pars
}
