% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gGamma.func}
\alias{gGamma.func}
\title{Calculate likelihood gradient in gamma}
\usage{
\code(gGamma.func(parList))
}
\description{
\code{gGamma.func} is a helper function to compute the
  likelihood gradient in gamma
}
