% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gGamma2.func}
\alias{gGamma2.func}
\title{Calculate likelihood gradient in gamma at update pars}
\usage{
\code{gGamma2.func(parList)}
}
\description{
\code{gGamma2.func} is a helper function to compute the
  likelihood gradient in gamma at update pars
}
