% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gPhi.func}
\alias{gPhi.func}
\title{Solve likelihood gradient given dispersion parameter}
\usage{
\code{gPhi.func(phi, yVec, muMat, tMat, yInd, LS)}
}
\description{
\code{gPhi.func} is a helper function to calculate the
  gradient of the likelihood of the observed data given the fitted
  mixture model for the purpose of finding the maximum likelihood
  dispersion parameter.
}
