% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gTilde.func}
\alias{gTilde.func}
\title{Calculate EM parameter updates}
\usage{
\code{gTilde.func(parList)}
}
\description{
\code{gTilde.func} is a helper function to compute the EM
  parameter update deltas
}
