% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{gTildeUpdate.func}
\alias{gTildeUpdate.func}
\title{Calculate EM parameter updates at update pars}
\usage{
\code(gTildeUpdate.func(parList))
}
\description{
\code{gTildeUpdate.func} is a helper function to compute the EM
  parameter update deltas at update pars
}
