% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{initEM.func}
\alias{initEM.func}
\title{Initialize EM steps}
\usage{
\code{initEM.func(y ,slope, subCDF, depth, stepInit)}
}
\description{
\code{initEM.func} is a helper function to compute the EM
  parameter updates
}
