% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{lLik.func}
\alias{lLik.func}
\title{Calculate EM log-likelihood}
\usage{
\code(lLik.func(parList))
}
\description{
\code{lLik.func} is a helper function to compute the EM
  log-likelihood up to an addative constant
}
