% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{lLikFull.func}
\alias{lLikFull.func}
\title{Calculate Full EM log-likelihood}
\usage{
\code{lLikFull.func(parList)}
}
\description{
\code{lLikFull.func} is a helper function to compute the EM
  log-likelihood up to an addative constant
}
