% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{lLikUpdate.func}
\alias{lLikUpdate.func}
\title{Calculate EM log-likelihood at update pars}
\usage{
\code{lLikUpdate.func(parList)}
}
\description{
\code{lLikUpdate.func} is a helper function to compute the EM
  log-likelihood up to an addative constant at update pars
}
