% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{lpMat.func}
\alias{lpMat.func}
\title{Calculate EM membership log-probabilities}
\usage{
\code(lpMat.func(parList))
}
\description{
\code{lpMat.func} is a helper function to compute the EM
  log membership probabilities
}
