% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{lpMatUpdate.func}
\alias{lpMatUpdate.func}
\title{Calculate EM membership log-probabilities at update pars}
\usage{
\code(lpMatUpdate.func(parList))
}
\description{
\code{lpMatUpdate.func} is a helper function to compute the EM
  log membership probabilities at update pars
}
