% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSimPlots.R
\docType{data}
\name{p100_multimodal}
\alias{p100_multimodal}
\title{Plot data from simulated expression}
\format{
Object of class \code{"gtable"}.
}
\usage{
data("p100_multimodal")
}
\description{
This plot is used in the vignette to demonstrate the flexibility of the
Dino model to smoothly estimate arbitrary latent multimodal expression
distributions
}
\examples{
data("p100_multimodal")
plot(p100_multimodal)

}
\keyword{datasets}
