% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSimPlots.R
\docType{data}
\name{p100_unimodal}
\alias{p100_unimodal}
\title{Plot data from simulated expression}
\format{
Object of class \code{"gtable"}.
}
\usage{
data("p100_unimodal")
}
\description{
This plot is used in the vignette to demonstrate the flexibility of the
Dino model to smoothly estimate arbitrary latent uninmodal expression
distributions
}
\examples{
data("p100_unimodal")
plot(p100_unimodal)

}
\keyword{datasets}
