% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{parGrad.func}
\alias{parGrad.func}
\title{Calculate EM parameter gradients}
\usage{
\code{parGrad.func(parList)}
}
\description{
\code{parGrad.func} is a helper function to compute the EM
  parameter gradients
}
