% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{parGradUpdate.func}
\alias{parGradUpdate.func}
\title{Calculate EM parameter gradients at update pars}
\usage{
\code{parGradUpdate.func(parList)}
}
\description{
\code{parGradUpdate.func} is a helper function to compute the EM
  parameter gradients at update pars
}
