% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{parListUpdate.func}
\alias{parListUpdate.func}
\title{Update parList by removing parameter}
\usage{
\code{parListUpdate.func(parList, doBeta = F)}
}
\description{
\code{parListUpdate.func} is a helper function to update
  parList when a parameter is to be removed
}
