% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSlope.R
\name{parPoisSlope}
\alias{parPoisSlope}
\title{Calculate parallel regression slopes}
\usage{
\code(parPoisSlope(datList, depth, clustVec, clustTab,
  prll, pctNZ = 0.01))
}
\description{
\code{parPoisSlope} is a helper function to compute poisson
  regression slopes
}
