% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{parResampCounts}
\alias{parResampCounts}
\title{Parallelize count resampling}
\usage{
\code{parResampCounts(countList, depth, depthRep,
  slope, prec, subInd, prll, emPar)}
}
\description{
\code{parResampCounts} is a helper function to parallelize the
  resampling algorithm.
}
