% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{parUpdate.func}
\alias{parUpdate.func}
\title{Calculate EM parameter updates}
\usage{
\code{parUpdate.func(parList)}
}
\description{
\code{parUpdate.func} is a helper function to compute the EM
  parameter updates
}
