% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmcSmall.R
\docType{data}
\name{pbmcSmall}
\alias{pbmcSmall}
\title{Subset of 500 peripheral blood mononuclear cells (PBMCs) from a healthy donor}
\format{
An object of class \code{"dgCMatrix"}.
}
\source{
\href{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}{3k PBMCs from a Healthy Donor}
}
\usage{
data("pbmcSmall")
}
\description{
This dataset derives from the "3k PBMCs from a Healthy Donor" public dataset
from 10X Genomics.
}
\examples{
data("pbmcSmall")
str(pbmcSmall)

}
\keyword{datasets}
