% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSlope.R
\name{poisSlope}
\alias{poisSlope}
\title{Calculate regression slopes}
\usage{
\code(poisSlope(y, depth, clustVec, clustTab, pctNZ = 0.01))
}
\description{
\code{poisSlope} is a helper function to compute poisson
  regression slopes
}
