% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{resampCounts}
\alias{resampCounts}
\title{Resample counts}
\usage{
\code{resampCounts(countMat, depth, slope, prec, subInd, emPar)}
}
\description{
\code{resampCounts} is a helper function to normalize counts
  by resampling from an estimated expression distribution conditional on
  the observed counts.
}
