% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dino.R
\name{SeuratFromDino}
\alias{SeuratFromDino}
\title{Create Seurat object from Dino normalized data}
\usage{
\code{SeuratFromDino(counts, doNorm = T, ...)}
}
\arguments{
\item{counts}{A numeric matrix of count data, either raw (eg. UMIs) or
normalized expression.}

\item{doNorm}{A logical indicating whether to normalize the input
\emph{counts} data before exporting results to a \emph{Seurat} object.
By default, it is assumed that the contents of \emph{counts} raw
expression which should be normalized.}

\item{doLog}{A logical indicating whether normalized counts should be log
transformed with a psuedocount of 1 prior to export.}

\item{...}{Additional parameters to pass to \code{Dino}}
}
\description{
\code{SeuratFromDino} is a wrapper simplifying the export of
  \code{Dino} normalized counts to a \emph{Seurat} object for secondary
  analysis.
}
