% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{solvePhi.func}
\alias{solvePhi.func}
\title{Find the maximum likelihood dispersion parameter}
\usage{
\code{solvePhi.func(yVec, muMat, bw, tMat, LS)}
}
\description{
\code{solvePhi.func} is a helper function to estimate the
  dispersion parameter of the observed data given the fitted mixture model.
}
