% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{splitGenes}
\alias{splitGenes}
\title{Split sparse matrix into list of columns}
\usage{
\code{splitGenes(counts, nCol = 40)}
}
\description{
\code{splitGenes} is a helper function to split a sparse count
  matrix into a list of gene sets for parallel computation.
}
