% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSimPlots.R
\docType{data}
\name{unimodalDat}
\alias{unimodalDat}
\title{Plot data from simulated expression}
\format{
Object of class \code{"gtable"}.
}
\usage{
data("unimodalDat")
}
\description{
This data is used in the vignette to demonstrate the flexibility of the
Dino model to smoothly estimate arbitrary latent uninmodal expression
distributions. These data are intended for internal use only.
}
\examples{
data("unimodalDat")

}
\keyword{datasets}
