% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEM.R
\name{zoom.func}
\alias{zoom.func}
\title{Find acceptable step length from bounds}
\usage{
\code(zoom.func(parList, alphaBounds))
}
\description{
\code{zoom.func} is a helper function to calculate the
  acceptable step length from bounds
}
