% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dino.R
\name{Dino}
\alias{Dino}
\title{Normalize scRNAseq data}
\usage{
\code{Dino(counts, nCores = 1, prec = 3, seed = 1, minNZ = 10,
  nSub = 1e4, depth = NULL, slope = NULL, returnMeta = F,
  emPar = list(maxIter = 100, tol = 1, conPar = 25), ...)}
}
\arguments{
\item{counts}{A numeric matrix object of expression counts - usually in
dgCMatrix format for memory efficiency. Column names denote cells
(samples or droplets) and row names denote genes.}

\item{nCores}{A non-negative integer scalar denoting the number of cores
which should be used. Setting nCores to 0 uses all cores as determined by
running \code{parallel::detectCores()}}

\item{prec}{A positive integer denoting the number of decimals to which to
round depth (if estimated internally via \code{depth = NULL})and
normalized counts for computational efficiency.}

\item{seed}{A numeric value denoting the random seed for replicability of
results. \code{Dino} automatically records the current random seed prior
to resetting to \emph{seed}, and resets to the recorded seed prior to
exiting.}

\item{minNZ}{A positive integer denoting the minimum number of non-zero
counts for a gene to be normalized by the Dino algorithm. It is
recommended to pre-filter the \emph{counts} matrix such that all genes
meet this threshold. Otherwise, genes with fewer than \emph{minNZ}
non-zeros will be scaled by depth for normalization.}

\item{nSub}{A positive integer denoting the number of samples to subset
for calculation of \emph{slope} and the EM algorithm.}

\item{depth}{A numeric vector of length equal to the columns of counts.
\emph{depth} denotes a median-centered, log-scale measure of cell-wise
sequencing depth. \code{Dino} defaults to defining depth as the
(within-cell) sum of counts across genes, followed by a log and
median-centering transformation.}

\item{slope}{A numeric scalar denoting the count-depth relationship on
the log-log scale. Typical values are close to 1 (implying a unit increase
in depth corresponds to a unit increase in expected counts on the log-log
scale), but may be higher, particularly in the case of non-UMI
protocols. \code{Dino} defaults to estimating \emph{slope} internally.}

\item{returnMeta}{A logical indicating whether metadata (sequencing depth
and slope) should be returned.}

\item{emPar}{A list of parameters to send to the EM algorithm.
\emph{maxIter} denotes the maximum number of model updates. \emph{tol}
denotes the cutoff threshold for reductions in the log likelihood
function. \emph{conPar} denotes the concentration parameter for the
resampling. \emph{conPar = 0} implies full resampling from the fitted
distribution. As \emph{conPar} increases, the normalized expression
converges to the scale-factor normalized values.}

\item{...}{Additional parameters to pass to \code{Scran::quickCluster}.}
}
\value{
\code{Dino} by default returns a matrix of normalized expression
  with identical dimensions as \emph{counts}. If \emph{returnMeta=T}, then
  \code{Dino} returns a list of normalized expression, sequencing depth, and
  slope.
}
\description{
\code{Dino} removes cell-to-cell variation in observed
counts due to the effects of sequencing depth from single-cell mRNA
sequencing experiments. \code{Dino} was particularly designed with UMI
based protocols in mind, but is applicable to non-UMI based chemistries
in the library preparation stage of sequencing.
}
